feat(api): require authorization to PATCH and DELETE event/

+++ backend/backend/views/event.py
@@ -4,6 +4,7 @@ from django.utils import timezone
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
 from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
 from backend.models import Event, VerificationToken
 from backend.serializers import EventSerializer
 
@@ -43,13 +44,25 @@ class EventAPIView(APIView):
         except:
             return JsonResponse({'error': 'Something went wrong'}, status=500)
 
-    def patch(self, request, pk):
+    def patch(self, request, identifier):
+        self.permission_classes = [IsAuthenticated]
+
+        authorization_token = request.headers.get('Authorization')
+        if not authorization_token:
+            return JsonResponse({'error': 'Unauthorized'}, status=401)
+
         try:
-            event = Event.objects.get(pk=pk)
+            if identifier.isdigit():
+                event = get_object_or_404(Event, pk=identifier)
+            else:
+                event = get_object_or_404(Event, event_slug=identifier)
         except Event.DoesNotExist:
             return JsonResponse({'error': 'Event not found'}, status=404)
         except:
             return JsonResponse({'error': 'Something went wrong'}, status=500)
+
+        if authorization_token != event.edit_token:
+            return JsonResponse({'error': 'Authorization token mismatch'}, status=401)
 
         try:
             serializer = EventSerializer(
@@ -62,9 +75,22 @@ class EventAPIView(APIView):
         except:
             return JsonResponse({'error': 'Something went wrong'}, status=500)
 
-    def delete(self, request, pk):
+    def delete(self, request, identifier):
+        self.permission_classes = [IsAuthenticated]
+
+        authorization_token = request.headers.get('Authorization')
+        if not authorization_token:
+            return JsonResponse({'error': 'Unauthorized'}, status=401)
+
         try:
-            event = Event.objects.get(pk=pk)
+            if identifier.isdigit():
+                event = get_object_or_404(Event, pk=identifier)
+            else:
+                event = get_object_or_404(Event, event_slug=identifier)
+
+            if authorization_token != event.edit_token:
+                return JsonResponse({'error': 'Authorization token mismatch'}, status=401)
+
             event.delete()
             return JsonResponse({}, status=204)
         except Event.DoesNotExist:
